+++ docs/changelog.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+* Expand :meth:`MessageFormat.format` to support ``date`` and ``datetime`` values.
+
 0.1.0 (2026-01-09)
 ------------------